+++ src/main/java/insane96mcp/survivalreimagined/module/movement/feature/TerrainSlowdown.java
@@ -2,6 +2,7 @@ package insane96mcp.survivalreimagined.module.movement.feature;
 
 import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
 import insane96mcp.insanelib.util.IdTagMatcher;
 import insane96mcp.insanelib.util.MCUtils;
@@ -11,9 +12,12 @@ import insane96mcp.survivalreimagined.module.Modules;
 import insane96mcp.survivalreimagined.module.movement.data.MaterialSlowdown;
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.Mth;
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Material;
 import net.minecraft.world.phys.AABB;
 import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.event.TickEvent;
@@ -49,6 +53,10 @@ public class TerrainSlowdown extends SRFeature {
 	public static ArrayList<MaterialSlowdown> materialOnSlowdown;
 	public static ArrayList<MaterialSlowdown> materialInSlowdown;
 
+	@Config
+	@Label(name = "Prevent Snow slowdown with Leather Boots")
+	public static Boolean preventSnowSlowdownWithLeatherBoots = true;
+
 	public TerrainSlowdown(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 	}
@@ -82,7 +90,6 @@ public class TerrainSlowdown extends SRFeature {
 	public void onPlayerTick(TickEvent.PlayerTickEvent event) {
 		if (!this.isEnabled()
 				|| event.phase != TickEvent.Phase.START
-				|| (event.player.tickCount + event.player.getId()) % 2 != 0
 				|| !event.player.isOnGround())
 			return;
 
@@ -114,6 +121,10 @@ public class TerrainSlowdown extends SRFeature {
 						}
 					}
 				}
+				if ((state.getMaterial() == Material.SNOW || state.getMaterial() == Material.TOP_SNOW)
+						&& event.player.getItemBySlot(EquipmentSlot.FEET).is(Items.LEATHER_BOOTS)
+						&& preventSnowSlowdownWithLeatherBoots)
+					continue;
 				onTerrainSlowdown += blockSlowdown;
 			}
 		}
@@ -136,6 +147,10 @@ public class TerrainSlowdown extends SRFeature {
 							break;
 						}
 					}
+					if ((state.getMaterial() == Material.SNOW || state.getMaterial() == Material.TOP_SNOW)
+							&& event.player.getItemBySlot(EquipmentSlot.FEET).is(Items.LEATHER_BOOTS)
+							&& preventSnowSlowdownWithLeatherBoots)
+						continue;
 					inTerrainSlowdown += blockSlowdown;
 				}
 			}